requirements.txt
langgraph
langchain-groq
yfinance
langchain-core
python-dotenv
streamlit
newsapi-python
langchain-tavily
langchain